refactor: split validation
