+++ caesar-cipher/my_caesar_cli.js
@@ -1,47 +1,12 @@
 const fs = require('fs');
 const {getTransformStream} = require('./caesarCipher');
-const arguments = process.argv.slice(2);
-
-
-
-const validateArguments = ({action, shift, inputPath, outputPath}) => {
-  if(!action || !shift) {
-    console.error('Action/Shift are required');
-    process.exit(-1);
-  }
-  
-  const isActionValid = action === 'encode' || action === 'decode';
-  const isShiftValid = Math.floor(Number(shift)) == shift;
-  
-  if(!isActionValid || !isShiftValid) {
-    console.error('Action or Shift have wrong value');
-    process.exit(-1);
-  }
-
-  if (inputPath) {
-    try {
-      fs.existsSync(inputPath);
-    } catch (e) {
-      process.stderr.write('Input file not found');
-    }
-  }
-
-  if (outputPath) {
-    try {
-      fs.existsSync(outputPath);
-    } catch (e) {
-      process.stderr.write('Output file not found');
-    }
-  }
-
-
-}
+const {action, shift, inputPath, outputPath} = require('./argument-parser');
 
 const outputData = ({sourcePath, action, shift, outputPath}) => {
   let writeStream = process.stdout;
 
   if(outputPath) {
-    writeStream = fs.createWriteStream(outputPath)
+    writeStream = fs.createWriteStream(outputPath, {flags:'a'})
   }
 
   let readStream = process.stdin;
@@ -55,17 +20,5 @@ const outputData = ({sourcePath, action, shift, outputPath}) => {
   readStream.pipe(transformStream).pipe(writeStream);
   
 }
-
-const getArgument = (args) => {
-  const argIndex = arguments.findIndex((el) => args.some((arg) => arg === el));
-  return  argIndex === -1 ? '' : arguments[argIndex + 1];
-}
-
-const action = getArgument(['-a', '--action']).toLowerCase();
-const shift = getArgument(['-s', '--shift']);
-const inputPath = getArgument(['-i', '--input']);
-const outputPath = getArgument(['-o', '--output']);
-
-validateArguments({action, shift, inputPath, outputPath});
 
 outputData({sourcePath: inputPath, action, shift, outputPath});
